drm/sde/rotator: Remove unneeded PM QoS requests
When the rotator is actually used (still an unsolved question in
computer science), these PM QoS requests block some CPUs in the LITTLE
cluster from entering deep idle because the driver assumes that display
rotating work occurs on a hardcoded set of CPUs, which is false. We
already have the IRQ PM QoS machinery for display rendering operations
that actually matter, so this cruft is unneeded.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I00c99b6754b011d5f92770d74483e287cc9e9731
Signed-off-by: pri0818 <priyanshusinghal0818@gmail.com>